refactor: Extract FeatureScheduler class
